requirements.txt
numpy==1.26.4
scipy==1.11.4
gradio
matplotlib
tqdm
opencv-python
scipy
einops
gdown
trimesh
pyglet<2
huggingface-hub[torch]>=0.22
evo
hydra-core
iopath
pillow
timm
ftfy
regex
typing_extensions
omegaconf
decord
pycocotools
transformers
accelerate
safetensors
tensorboard
jupyter
notebook
ipywidgets
ipycanvas
ipympl
scikit-image
scikit-learn
h5py
mmcv-lite
mmengine
lightning
pytorch-lightning
torchmetrics
nuscenes-devkit
roma
pycolmap
viser
supervision
pyyaml
scipy
filelock
idr-torch